{"types":"C","runs":5}

+++ problem15.java
@@ -1,8 +1,8 @@
 import acm.program.*;
 public class problem15 extends ConsoleProgram {
 	public void run() {
-		int n1 = readInt("first number:");
-		int n2 = readInt("second number:");
+		double n1 = readDouble("first number:");
+		double n2 = readDouble("second number:");
 		println((n1 + n2)/2);
 	}
 }
